requirements.txt
### Dsicord
discord.py

### Other
transformers
accelerate
diffusers
googletrans==4.0.0rc1
python-dotenv